feat: add cache(drift) & implement provider in bar graph

+++ lib/domain/entity/steps/step_summary_entity.dart
@@ -3,7 +3,7 @@ import 'package:equatable/equatable.dart';
 class StepSummaryEntity  extends Equatable{
   final int totalSteps;
   final int totalGoal;
-  final double avgGoal;
+  final int avgGoal;
 
   const StepSummaryEntity({
     required this.totalSteps,
